+++ java-library/src/main/java/com/microsoft/azure/functions/openai/annotation/assistant/AssistantPost.java
@@ -130,4 +130,11 @@ public @interface AssistantPost {
      * @return The maxTokens value.
      */
     String maxTokens() default "100";
+
+    /**
+     * Indicates whether the assistant uses a reasoning model.
+     *
+     * @return {@code true} if the assistant is based on a reasoning model; {@code false} otherwise.
+     */
+    boolean isReasoningModel() default false;
 }