+++ java-library/src/main/java/com/microsoft/azure/functions/openai/annotation/search/SemanticSearch.java
@@ -138,4 +138,11 @@ public @interface SemanticSearch {
      * @return The maxTokens value.
      */
     String maxTokens() default "2048";
+
+    /**
+     * Indicates whether the assistant uses a reasoning model.
+     *
+     * @return {@code true} if the assistant is based on a reasoning model; {@code false} otherwise.
+     */
+    boolean isReasoningModel();
 }